# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "天气管理系统后端API文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9027",
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "用户输入用户名和密码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户对象，包括用户名和密码两个属性",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    },
                    "422": {
                        "description": "密码错误",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    },
                    "442": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    },
                    "500": {
                        "description": "系统发放token异常",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "用于管理员注册新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "新用户，传入用户名和密码",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "442": {
                        "description": "用户已存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取登录用户的信息",
                "responses": {
                    "200": {
                        "description": "返回用户id，头像avatar，用户名name,用户角色role",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "管理员修改用户信息",
                "parameters": [
                    {
                        "description": "新用户的用户名、密码",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "具体修改的用户的ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "管理员添加用户",
                "parameters": [
                    {
                        "description": "新用户的用户名、密码",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "442": {
                        "description": "用户已存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "管理员修改用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "具体修改的用户的ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
        "/user/users": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "管理员获取所有用户信息及其设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页数，默认为1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数据个数，默认为5",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回用户列表users，用户个数count",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.SwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "msg": {
                    "description": "描述",
                    "type": "string"
                }
            }
        },
        "controller.SwaggerResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "msg": {
                    "description": "描述",
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "description": "0:管理员 , 1: 普通用户",
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                }
            }
        }
    }
}